Move node pool options func to vke.go
